requirements.txt
absl-py==0.13.0
APScheduler==3.6.3
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.4
clang==5.0
cycler==0.10.0
flatbuffers==1.12
gast==0.4.0
grpcio==1.39.0
h11==0.9.0
h2==3.2.0
h5py==3.1.0
hpack==3.0.0
hstspreload==2021.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.3
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.2
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.15.0
sklearn==0.0
sniffio==1.2.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==2.2.0
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==3.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
